Category/Precision Measurement & Mapping/2021D/requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
PyYAML>=5.4.0
requests>=2.25.0
